feat-#16: move entities to infrastructure
