+++ src/database/config/typeOrmAsyncConfig.ts
@@ -1,7 +1,7 @@
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { TypeOrmModuleAsyncOptions } from '@nestjs/typeorm';
 
-import { ItemEntity } from '../../item/entities';
+import { ItemEntity } from '../../item/infrastructure/entities';
 
 export const typeOrmAsyncConfig: TypeOrmModuleAsyncOptions = {
   imports: [ConfigModule],
