+++ src/item/application/commands/handlers/create-item.handler.ts
@@ -2,7 +2,7 @@ import { CommandHandler, EventPublisher, ICommandHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { ErrorValidationService, RpcExceptionService } from '../../../../utils';
-import { ItemEntity } from '../../../entities';
+import { ItemEntity } from '../../../infrastructure/entities';
 import { ItemRepository } from '../../../repositories';
 import { ItemWriteRepository } from '../../../infrastructure/repositories';
 import { CreateItemCommand } from '../impl';
