+++ src/item/application/commands/handlers/delete-item-by-id.handler.ts
@@ -3,7 +3,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 
 import { DeleteItemByIdCommand } from '../impl';
 import { RpcExceptionService, ErrorValidationService } from '../../../../utils';
-import { ItemEntity } from '../../../entities';
+import { ItemEntity } from '../../../infrastructure/entities';
 import { ItemRepository } from '../../../repositories';
 import { ItemWriteRepository } from '../../../infrastructure/repositories';
