+++ src/item/infrastructure/entities/item.entity.ts
